+++ agb/src/display/object.rs
@@ -794,7 +794,7 @@ impl ObjectController {
         });
 
         let loan = Loan {
-            index: index as u8,
+            index,
             phantom: PhantomData,
         };
 
@@ -916,8 +916,8 @@ impl<'a> Object<'a> {
     /// [ObjectController::commit] is called.
     pub fn set_x(&mut self, x: u16) -> &mut Self {
         let object_inner = unsafe { self.object_inner() };
-        object_inner.attrs.a1a.set_x(x.rem_euclid(1 << 9) as u16);
-        object_inner.attrs.a1s.set_x(x.rem_euclid(1 << 9) as u16);
+        object_inner.attrs.a1a.set_x(x.rem_euclid(1 << 9));
+        object_inner.attrs.a1s.set_x(x.rem_euclid(1 << 9));
         self
     }